YH-406: Refactored TextHtml component
- Wrap all content in <pre> tag
- Style code snippets
- Move component to shared/ui directory
