+++ src/shared/ui/TextHtml/index.ts
@@ -1,0 +1,1 @@
+export { TextHtml } from './ui/TextHtml';
