+++ src/shared/ui/TextHtml/ui/TextHtml.module.css
@@ -1,0 +1,5 @@
+.text code {
+  padding: 5px;
+  border-radius: 12px;
+  background-color: var(--background-nav-item);
+}
